refactor: add index.ts for `deployments/src`
