+++ pkg/deployments/src/index.ts
@@ -1,0 +1,6 @@
+export * from './forkTests';
+export * from './signers';
+export * from './task';
+export * from './test';
+
+export { default as Task } from './task';
