+++ pkg/deployments/tasks/20220916-batch-relayer-v4/test/test.fork.ts
@@ -4,15 +4,13 @@ import { BigNumber, Contract } from 'ethers';
 
 import { BigNumberish, fp } from '@balancer-labs/v2-helpers/src/numbers';
 
-import { describeForkTest } from '../../../src/forkTests';
-import Task, { TaskMode } from '../../../src/task';
-import { getForkedNetwork } from '../../../src/test';
-import { impersonate } from '../../../src/signers';
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers';
 import { WeightedPoolEncoder } from '@balancer-labs/balancer-js';
 
 import { MAX_UINT256 } from '@balancer-labs/v2-helpers/src/constants';
 import { defaultAbiCoder } from '@ethersproject/abi/lib/abi-coder';
+
+import { describeForkTest, impersonate, getForkedNetwork, Task, TaskMode } from '../../../src';
 
 describeForkTest('BatchRelayerLibrary', 'mainnet', 15485000, function () {
   let task: Task;
