+++ pkg/deployments/tasks/deprecated/20210721-liquidity-bootstrapping-pool/test/task.fork.ts
@@ -11,10 +11,7 @@ import { calculateInvariant } from '@balancer-labs/v2-helpers/src/models/pools/w
 import { expectEqualWithError } from '@balancer-labs/v2-helpers/src/test/relativeError';
 import { advanceToTimestamp, currentTimestamp, DAY, MINUTE, MONTH } from '@balancer-labs/v2-helpers/src/time';
 
-import { describeForkTest } from '../../../../src/forkTests';
-import Task, { TaskMode } from '../../../../src/task';
-import { getForkedNetwork } from '../../../../src/test';
-import { getSigner, impersonateWhale } from '../../../../src/signers';
+import { describeForkTest, getSigner, impersonateWhale, getForkedNetwork, Task, TaskMode } from '../../../../src';
 
 describeForkTest('LiquidityBootstrappingPoolFactory', 'mainnet', 14850000, function () {
   let owner: SignerWithAddress, whale: SignerWithAddress;
