+++ pkg/deployments/tasks/deprecated/20210727-meta-stable-pool/test/task.fork.ts
@@ -10,10 +10,7 @@ import { expectEqualWithError } from '@balancer-labs/v2-helpers/src/test/relativ
 import { MAX_UINT256, ZERO_ADDRESS } from '@balancer-labs/v2-helpers/src/constants';
 import { StablePoolEncoder, SwapKind } from '@balancer-labs/balancer-js';
 
-import { describeForkTest } from '../../../../src/forkTests';
-import Task, { TaskMode } from '../../../../src/task';
-import { getForkedNetwork } from '../../../../src/test';
-import { getSigner, impersonateWhale } from '../../../../src/signers';
+import { describeForkTest, getSigner, impersonateWhale, getForkedNetwork, Task, TaskMode } from '../../../../src';
 
 describeForkTest('MetaStablePoolFactory', 'mainnet', 14850000, function () {
   let owner: SignerWithAddress, whale: SignerWithAddress;
