+++ pkg/deployments/tasks/deprecated/20210907-investment-pool/test/task.fork.ts
@@ -11,10 +11,7 @@ import { calculateInvariant } from '@balancer-labs/v2-helpers/src/models/pools/w
 import { expectEqualWithError } from '@balancer-labs/v2-helpers/src/test/relativeError';
 import { advanceToTimestamp, currentTimestamp, DAY, MINUTE, MONTH } from '@balancer-labs/v2-helpers/src/time';
 
-import { describeForkTest } from '../../../../src/forkTests';
-import Task, { TaskMode } from '../../../../src/task';
-import { getForkedNetwork } from '../../../../src/test';
-import { getSigners, impersonateWhale } from '../../../../src/signers';
+import { describeForkTest, getSigners, impersonateWhale, getForkedNetwork, Task, TaskMode } from '../../../../src';
 
 describeForkTest('InvestmentPoolFactory', 'mainnet', 14850000, function () {
   let owner: SignerWithAddress, wallet: SignerWithAddress, whale: SignerWithAddress;
